satellite/requirements.txt
affine==2.4.0
attrs==25.4.0
blinker==1.9.0
cachetools==7.0.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cligj==0.7.2
cloudpickle==3.1.2
colorama==0.4.6

# ---- CORE COMPATIBILITY PINS ----
numpy==1.26.4
rasterio==1.4.4
dask==2024.8.0
distributed==2024.8.0
# --------------------------------

Flask==3.0.0
Flask-Cors==4.0.0
fsspec==2026.1.0
geopandas==1.1.2
idna==3.11
imageio==2.37.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy_loader==0.4
locket==1.0.0
MarkupSafe==3.0.3
msgpack==1.1.2
networkx==3.6.1

odc-geo==0.5.0
odc-stac==0.3.9

packaging==26.0
pandas==3.0.0
partd==1.4.2
pillow==12.1.0
psutil==7.2.2
pyogrio==0.12.1
pyparsing==3.3.2
pyproj==3.7.2
pystac==1.14.3
pystac-client==0.7.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
scikit-image==0.26.0
scipy==1.17.0
shapely==2.1.2
six==1.17.0
sortedcontainers==2.4.0
tblib==3.2.2
tifffile==2026.1.28
toolz==1.1.0
tornado==6.5.4
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.5
xarray==2026.1.0
zict==3.0.0
gunicorn==21.2.0